Integrate all tests into unit testing.
There are no external tests now.

Also fix the debug formatting for Credential Stores.
